drop va_list from formals if requested
I'm working on a feature that involves creating wrappers for some
functions, using alternate means to pass variable argument lists to
the wrapped versions.  The split-out (wrapped) function shouldn't be
stdarg, and though comments in m_always_copy_start in
ipa_param_adjustments suggested that making it negative would cause
variable arguments to be dropped, this was not the case, and AFAICT
there was no way to accomplish that.

Perhaps there was no such intent implied by the comments, but that
sounded like a reasonable plan to me, so I went ahead an implemented
it.  With this patch, a negative m_always_copy_start drops the
variable argument list marker from a cloned function's type, and the
stdarg flag from the cloned cfun data structure.


for  gcc/ChangeLog

	* ipa-param-manipulation.c (build_adjusted_function_type): Add
	skip_valist, obey it.
	(ipa_param_adjustments::build_new_function_type): Take
	negative m_always_copy_start as skip_valist.
	(ipa_param_body_adjustmnets::modify_formal_parameters):
	Likewise.
	* tree-inline.c (initialize_cfun): Clear stdarg if function
	type doesn't support it.
	* ipa-strub.c (pass_ipa_strub::execute): Check that stdarg is
	dropped from the clone.